refactor: rename services and update import paths for better clarity
