+++ package.json
@@ -13,6 +13,7 @@
     "@types/react-dom": "^18.3.1",
     "axios": "^1.7.7",
     "chart.js": "^4.4.6",
+    "date-fns": "^4.1.0",
     "dotenv": "^16.4.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
