+++ src/interfaces/Interfaces.tsx
@@ -11,4 +11,11 @@ export interface Issue {
 export interface Milestone {
     number: number;
     title: string;
+    closed_at: string;
+    closed_issues: number;
+    state: string;
+    html_url: string;
+    due_on: string;
+    start_on: string;
+    days: string[];
 }
